Recommended changes to Conviction documentation to allow copy-paste ru (#92)
NOTE: Needs confirmation that my changes match intent of documentation
in terms of the proper selection of details that are provided in the
example and surrounding context. Several things have changed since this
documentation was first drafted and may need revised. Will take any
direction on changing of wording and/or direct dictation.

Changes to examples to fix syntax issues and allow copy-paste runs.
Tested in python 3.9 and 3.13 in conda environments.